web app fixes and cleanup
renamed bitstream and included an empty fpga.bit to remain compatible with
future changes to application startup. we use a special loading procedure, so
we better not get mixed up with the generic one.

removed pipe functions and corresponding options, because our memory seems to
not be eligible for splicing.

fixes pthread usage and resource handling in web app worker.

please let's use tabs for block indentation only, and spaces for multi-line
alignment. it makes differing tab-widths unobtrusive.